requirements.txt
tensorflow
matplotlib
numpy
Pillow
flask